chore: 🔧 move share button to a shared component
